+++ internal/backend/remote-state/s3/validate.go
@@ -11,6 +11,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/aws/arn"
 	"github.com/hashicorp/terraform/internal/tfdiags"
 	"github.com/zclconf/go-cty/cty"
@@ -116,6 +117,37 @@ func validateStringMatches(re *regexp.Regexp, description string) stringValidato
 				path,
 			))
 		}
+	}
+}
+
+func validateStringInSlice(sl []string) stringValidator {
+	return func(val string, path cty.Path, diags *tfdiags.Diagnostics) {
+		match := false
+		for _, s := range sl {
+			if val == s {
+				match = true
+			}
+		}
+		if !match {
+			*diags = diags.Append(attributeErrDiag(
+				"Invalid Value",
+				fmt.Sprintf("Value must be one of [%s]", strings.Join(sl, ", ")),
+				path,
+			))
+		}
+
+	}
+}
+
+// validateStringRetryMode ensures the provided value in a valid AWS retry mode
+func validateStringRetryMode(val string, path cty.Path, diags *tfdiags.Diagnostics) {
+	_, err := aws.ParseRetryMode(val)
+	if err != nil {
+		*diags = diags.Append(attributeErrDiag(
+			"Invalid Value",
+			err.Error(),
+			path,
+		))
 	}
 }